requirements.txt
# Requirements for Gaze Tracking Service
flask==2.3.3
flask-cors==4.0.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
gaze-tracking==0.0.1